Add displayingResult to main.js
The displayingResult boolean helps tell the
difference between the display having data
from a previous calculation (true) vs when the
display shows the calculator receiving data.
This is so after = is pressed, the display clears
for the new input, yet won't continue erasing
and adding the next digit for numberA. This is
more like object disoriented programming.