Add logic for chaining operators
enoughOperands function checks if calculate
function should run. In a regular A-op-B-eq
sequence, the operatorButton listener ran as
before.

If the sequence is A-op1-B-op2,
the operatorButton listener upon hearing op2
clears the display,
runs calculate on the A-op1-B,
and displays the result.
Then it parses this result
as numberA, and changeOperator to the current
operator (op2).